+++ src/Typesense.php
@@ -43,7 +43,7 @@ use yii\base\Event;
 
 /**
  * Craft plugins are very much like little applications in and of themselves. We’ve made
- * it as simple as we can, but the training wheels are off. A little prior knowledge is
+ * it as simple as we can, but the training wheels are off. Little prior knowledge is
  * going to be required to write a plugin.
  *
  * For the purposes of the plugin docs, we’re going to assume that you know PHP and SQL,
@@ -60,6 +60,7 @@ use yii\base\Event;
  * @property  SynonymService $synonymService
  *
  * @property  Settings $settings
+ * @property mixed|object|null $synonyms
  */
 class Typesense extends Plugin
 {